test: try a real live test
